Move combo edit form to ComboList, add GameName
The problem with appending a combined create/edit combo form is that
when a combo that is not the last in the list, the form gets added to
the end of the list, instead of replacing the to-be-edited combo. By
moving the combo edit form to within the ComboList form, we can render
the form in place of the target combo.